change ssize_t to long to solve compile error

+++ leetcode/binarySearch.h
@@ -1,10 +1,10 @@
 #include <vector>
 template <class T>
-ssize_t BinarySearchWithWhile(std::vector<T> &container, T target) {
-  ssize_t left = 0;
-  ssize_t right = container.size() - 1;
+long BinarySearchWithWhile(std::vector<T> &container, T target) {
+  long left = 0;
+  long right = container.size() - 1;
   while (left <= right) {
-    ssize_t mid = (left + right) / 2;
+    long mid = (left + right) / 2;
     if (container[mid] == target) {
       return mid;
     }
